+++ src/main-curriculum/components/custom-aside.tsx
@@ -2,6 +2,8 @@ import { IoPerson } from "react-icons/io5";
 
 import type { PersonalInfoResponse } from "../interfaces";
 import { getCandidateContactInfoMapped } from "../helpers";
+import { getIconFromIconKey } from "../helpers/icons";
+import { getCurriculumPdf } from "../endpoints";
 interface Props {
   personalInfo: PersonalInfoResponse;
 }
@@ -10,12 +12,23 @@ const BASE_URL = import.meta.env.VITE_BASE_URL;
 
 export const CustomAside = ({ personalInfo }: Props) => {
   const canditateContactInfo = getCandidateContactInfoMapped(personalInfo);
+  const handleClick = async () => {
+    const response = await getCurriculumPdf();
+
+    const url = window.URL.createObjectURL(new Blob([response]));
+    const link = document.createElement("a");
+    link.href = url;
+    link.setAttribute("download", "curriculum.pdf");
+    document.body.appendChild(link);
+    link.click();
+    link.remove();
+  };
 
   return (
     <aside className="bg-orange-50 p-4 flex flex-col items-center gap-6 h-full">
       <h2 className="text-blue-700 text-xl underline md:self-start flex items-center gap-1">
         <IoPerson />
-        <span>Datos de contacto:</span>
+        <span>Datos de contacto</span>
       </h2>
       <img
         className="rounded-xl w-2/4 md:w-1/2"
@@ -35,6 +48,14 @@ export const CustomAside = ({ personalInfo }: Props) => {
           );
         })}
       </ul>
+
+      <button
+        className="p-2 flex gap-2 justify-center items-center w-full rounded-xl cursor-pointer bg-sky-600 hover:bg-sky-500 text-white"
+        onClick={handleClick}
+      >
+        <span>{getIconFromIconKey("downloadCV")}</span>
+        <span>Descarga Curriculum</span>
+      </button>
     </aside>
   );
 };