+++ src/main-curriculum/components/custom-main/custom-main.tsx
@@ -1,9 +1,11 @@
 import { useSection } from "../../hooks";
 import { SectionItem } from "./section-item";
+import { SoftSkillsSection } from "./soft-skills-section";
 
 export const CustomMain = () => {
   const { sectionsData: aboutMeSectionData } = useSection("about-me");
   const { sectionsData: restOfSections } = useSection("rest-of-sections");
+  const { sectionsData: softSkillsSection } = useSection("soft-skills");
 
   return (
     <>
@@ -11,14 +13,19 @@ export const CustomMain = () => {
         <SectionItem curriculumItem={aboutMeSectionData[0]} />
       )}
 
-      {restOfSections?.map((sectionItemData) => {
-        return (
-          <SectionItem
-            key={sectionItemData.title}
-            curriculumItem={sectionItemData}
-          />
-        );
-      })}
+      {restOfSections &&
+        restOfSections?.map((sectionItemData) => {
+          return (
+            <SectionItem
+              key={sectionItemData.title}
+              curriculumItem={sectionItemData}
+            />
+          );
+        })}
+
+      {softSkillsSection && (
+        <SoftSkillsSection softSkillsSection={softSkillsSection} />
+      )}
     </>
   );
 };