+++ src/main-curriculum/components/custom-main/index.ts
@@ -1,2 +1,3 @@
 export * from "./custom-main";
 export * from "./section-item";
+export * from "./soft-skills-section";