+++ src/main-curriculum/endpoints/index.ts
@@ -2,3 +2,5 @@ export * from "./axios-instance";
 export * from "./personal-info";
 export * from "./about-me";
 export * from "./sections";
+export * from "./soft-skills";
+export * from "./curriculum-pdf";