+++ src/main-curriculum/helpers/icons.tsx
@@ -3,12 +3,16 @@ import { BsFillPersonBadgeFill } from "react-icons/bs";
 import { IoSchoolSharp } from "react-icons/io5";
 import { FaBriefcase } from "react-icons/fa";
 import { MdVolunteerActivism } from "react-icons/md";
+import { GiSkills } from "react-icons/gi";
+import { FaFilePdf } from "react-icons/fa6";
 
 export const icons: Record<string, IconType> = {
   personBadge: BsFillPersonBadgeFill,
   academicTraining: IoSchoolSharp,
   workExperience: FaBriefcase,
   volunteering: MdVolunteerActivism,
+  softSkills: GiSkills,
+  downloadCV: FaFilePdf,
 };
 
 export const getIconFromIconKey = (iconKey: string) => {