+++ src/main-curriculum/hooks/use-section.ts
@@ -1,7 +1,11 @@
 import { useEffect, useState } from "react";
 
 import type { CurriculumSectionResponse } from "../interfaces";
-import { getAboutMeSection, getSections } from "../endpoints";
+import {
+  getAboutMeSection,
+  getSections,
+  getSoftSkillsSection,
+} from "../endpoints";
 
 export const useSection = (sectionName: string) => {
   const [sectionsData, setSectionsData] = useState<
@@ -9,8 +13,18 @@ export const useSection = (sectionName: string) => {
   >(null);
 
   useEffect(() => {
-    const endpoints =
-      sectionName === "about-me" ? getAboutMeSection : getSections;
+    let endpoints = getSections;
+
+    switch (sectionName) {
+      case "about-me":
+        endpoints = getAboutMeSection;
+        break;
+      case "soft-skills":
+        endpoints = getSoftSkillsSection;
+        break;
+      default:
+        endpoints = getSections;
+    }
 
     endpoints().then((sections) => {
       setSectionsData([...sections]);