+++ CHANGELOG.md
@@ -5,6 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2025-10-20
+
+### Changed
+
+- Refactor useSection to integrate soft-skills endpoint.
+- Extract logic from custo main related to soft-skills to a functional component.
+- Change badge color.
+
+### Added
+
+- Added more icons at icons catalog.
+- Added getCurriculumPdf endpoint.
+- Added curriculum download button to aside.
+- Added getSoftSkillsSection endpoint.
+- Added softSkills section to custom main component.
+
 ## [1.0.0] - 2025-10-20
 
 ### Changed